+++ client/lib/ft4/crosschain/index.ts
@@ -1,0 +1,24 @@
+// export the needed stuff
+
+export {
+  isTransferApplied,
+  getPendingTransfersForAccount,
+  getAssetOriginById,
+} from "./query-functions";
+export { createOrchestrator } from "./orchestrator";
+export {
+  initTransfer,
+  applyTransfer,
+  getInitTransferArgs,
+} from "./op-functions";
+export { findPathToChainForAsset, PathfinderError } from "./pathfinder";
+
+export { pendingTransfersForAccount } from "./queries";
+
+export {
+  Orchestrator,
+  OrchestratorEvents,
+  GtvInitTransferArgs,
+  PendingTransfer,
+  PendingTransferResponse,
+} from "./types";