+++ client/lib/ft4/ft-session.ts
@@ -155,7 +155,7 @@ export function createAuthDataService(connection: Connection): AuthDataService {
     },
     getNonce: async (accountId: BufferId, authDescriptorId: BufferId) =>
       connection.query(nonce(accountId, authDescriptorId)),
-    getLoginConfig: async (configName: string | null = null) =>
+    getLoginConfig: async (configName: string | undefined = undefined) =>
       connection.query(loginConfig(configName)),
     getBrid: () => Buffer.from(connection.client.config.blockchainRid, "hex"),
   });
@@ -187,7 +187,7 @@ export function createKeyStoreInteractor(
     getLoginManager: (loginKeyStore?: LoginKeyStore) =>
       createLoginManager(connection, keyStore, loginKeyStore),
     onKeyStoreChanged: async (handler: (arg0: KeyStoreInteractor) => void) => {
-      ftEventEmitter.on("KeyStoreChanged", (newKeyStore: KeyStore) =>
+      ftEventEmitter.on("KeyStoreChange", (newKeyStore: KeyStore) =>
         handler(createKeyStoreInteractor(client, newKeyStore)),
       );
     },